basic-with-postgresql/pyproject.toml
[project]
name = "basic-with-postgresql"
version = "0.1.0"
description = "minimum application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular[sidecar]>=0.29.0",
    "logger>=1.4",
    "markdown>=3.10",
    "psycopg2>=2.9.11",
    "pytest-django>=4.11.1",
    "ruff>=0.14.10",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.pytest.ini_options]
addopts = "-ra -q --reuse-db"
DJANGO_SETTINGS_MODULE = "src.settings"
pythonpath = "."
testpaths = ["tests"]
